+++ space-engineers-utilities/seut_window_manager.py
@@ -134,18 +134,11 @@ class SEUT_WindowManager(PropertyGroup):
     bboxColor: FloatVectorProperty(
         name="Color",
         description="The color of the Bounding Box",
-        subtype='COLOR',
-        default=(0.42, 0.827, 1),
-        update=update_BBox
-    )
-    bboxTransparency: FloatProperty(
-        name="Transparency",
-        description="The transparency of the Bounding Box",
-        default=0.3,
-        max=1,
-        min=0,
-        step=10.0,
-        update=update_BBox
+        subtype='COLOR_GAMMA',
+        size=4,
+        min=0.0,
+        max=1.0,
+        default=(0.42, 0.827, 1, 0.3)
     )
 
     fix_scratched_materials: BoolProperty(