+++ Changelog.md
@@ -1,6 +1,7 @@
 # Changelog
 * Added: Support for using Better FBX importer code in SEUT import.
 * Added: Import materials from model XMLs.
+* Added: Ability to customize bounding box color and transparency.
 * Improved: Code around creation of subpart empties. Now handles targeted objects better.
 * Added: Automatic material import from model XMLs during FBX import (only materials defined in that XML, not MaterialRefs).
 * Added: Warning for main- and subpart-scene having different `Grid Export`-settings.